complete page chart

+++ admin/View/quanLySP.php
@@ -12,12 +12,10 @@ include 'View/titleOfComponents.php';
         <div class="card">
           <div class="card-header d-flex justify-content-between align-items-center">
             <h5>Danh Mục</h5>
-            <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal"
-              data-whatever="">Thêm mới
+            <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal" data-whatever="">Thêm mới
             </button>
 
-            <div class="modal fade" id="exampleModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel"
-              aria-hidden="true">
+            <div class="modal fade" id="exampleModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
               <div class="modal-dialog" role="document">
                 <div class="modal-content">
                   <div class="modal-header">
@@ -34,8 +32,7 @@ include 'View/titleOfComponents.php';
                       </div>
                       <div class="form-group">
                         <label for="ID_group" class="col-form-label">ID_group</label>
-                        <input type="text" class="form-control" id="ID_group" name="id_group"
-                          placeholder="yêu cầu viết liền">
+                        <input type="text" class="form-control" id="ID_group" name="id_group" placeholder="yêu cầu viết liền">
                       </div>
                       <button class="btn btn-success">Thêm</button>
                     </form>
@@ -57,35 +54,28 @@ include 'View/titleOfComponents.php';
                 </thead>
                 <tbody>
                   <?php foreach ($allDanhMuc as $danhmuc) : ?>
-                  <tr>
-                    <th scope="row"><?= $danhmuc["id"] ?></th>
-                    <td><?= $danhmuc["name"] ?></td>
-                    <td><?= $danhmuc["id_group"] ?></td>
-                    <td class="box-active">
-                      <a href="?act=deleteDm&id=<?= $danhmuc["id"] ?>"
-                        onclick=" return confirm('Bạn có chắc muốn xóa không')"><button type="button"
-                          class="btn rounded btn-primary">
-                          <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor"
-                            class="bi bi-trash" viewBox="0 0 16 16">
-                            <path
-                              d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6Z">
-                            </path>
-                            <path
-                              d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1ZM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118ZM2.5 3h11V2h-11v1Z">
+                    <tr>
+                      <th scope="row"><?= $danhmuc["id"] ?></th>
+                      <td><?= $danhmuc["name"] ?></td>
+                      <td><?= $danhmuc["id_group"] ?></td>
+                      <td class="box-active">
+                        <a href="?act=deleteDm&id=<?= $danhmuc["id"] ?>" onclick=" return confirm('Bạn có chắc muốn xóa không')"><button type="button" class="btn rounded btn-primary">
+                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash" viewBox="0 0 16 16">
+                              <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6Z">
+                              </path>
+                              <path d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1ZM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118ZM2.5 3h11V2h-11v1Z">
+                              </path>
+                            </svg>
+                          </button>
+                        </a>
+                        <a href="?act=editDm&id=<?= $danhmuc["id"] ?>" class="btn btn-success rounded">
+                          <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-pen" viewBox="0 0 16 16">
+                            <path d="m13.498.795.149-.149a1.207 1.207 0 1 1 1.707 1.708l-.149.148a1.5 1.5 0 0 1-.059 2.059L4.854 14.854a.5.5 0 0 1-.233.131l-4 1a.5.5 0 0 1-.606-.606l1-4a.5.5 0 0 1 .131-.232l9.642-9.642a.5.5 0 0 0-.642.056L6.854 4.854a.5.5 0 1 1-.708-.708L9.44.854A1.5 1.5 0 0 1 11.5.796a1.5 1.5 0 0 1 1.998-.001zm-.644.766a.5.5 0 0 0-.707 0L1.95 11.756l-.764 3.057 3.057-.764L14.44 3.854a.5.5 0 0 0 0-.708l-1.585-1.585z">
                             </path>
                           </svg>
-                        </button>
-                      </a>
-                      <a href="?act=editDm&id=<?= $danhmuc["id"] ?>" class="btn btn-success rounded">
-                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor"
-                          class="bi bi-pen" viewBox="0 0 16 16">
-                          <path
-                            d="m13.498.795.149-.149a1.207 1.207 0 1 1 1.707 1.708l-.149.148a1.5 1.5 0 0 1-.059 2.059L4.854 14.854a.5.5 0 0 1-.233.131l-4 1a.5.5 0 0 1-.606-.606l1-4a.5.5 0 0 1 .131-.232l9.642-9.642a.5.5 0 0 0-.642.056L6.854 4.854a.5.5 0 1 1-.708-.708L9.44.854A1.5 1.5 0 0 1 11.5.796a1.5 1.5 0 0 1 1.998-.001zm-.644.766a.5.5 0 0 0-.707 0L1.95 11.756l-.764 3.057 3.057-.764L14.44 3.854a.5.5 0 0 0 0-.708l-1.585-1.585z">
-                          </path>
-                        </svg>
-                      </a>
-                    </td>
-                  </tr>
+                        </a>
+                      </td>
+                    </tr>
                   <?php endforeach; ?>
                 </tbody>
               </table>
@@ -97,11 +87,9 @@ include 'View/titleOfComponents.php';
         <div class="card">
           <div class="card-header d-flex justify-content-between align-items-center">
             <h5>Danh Sách Món Ăn</h5>
-            <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal1"
-              data-whatever="">Thêm mới
+            <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal1" data-whatever="">Thêm mới
             </button>
-            <div class="modal fade" id="exampleModal1" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel"
-              aria-hidden="true">
+            <div class="modal fade" id="exampleModal1" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
               <div class="modal-dialog" role="document">
                 <div class="modal-content">
                   <div class="modal-header">
@@ -117,14 +105,10 @@ include 'View/titleOfComponents.php';
                         <input name="name" required type="text" class="form-control">
                       </div>
                       <div class="form-group">
-                        <label class="col-form-label">Tên Cú Pháp:</label>
-                        <input name="referred" required type="text" class="form-control">
-                      </div>
-                      <div class="form-group">
                         <label class="col-form-label">Danh Mục:</label>
                         <select class="form-control" name="id_dm">
                           <?php foreach ($allDanhMuc as $danhmuc) : ?>
-                          <option value="<?= $danhmuc["id"] ?>"><?= $danhmuc["name"] ?></option>
+                            <option value="<?= $danhmuc["id"] ?>"><?= $danhmuc["name"] ?></option>
                           <?php endforeach; ?>
                         </select>
                       </div>
@@ -159,37 +143,30 @@ include 'View/titleOfComponents.php';
                 </thead>
                 <tbody>
                   <?php foreach ($allFood as $food) : ?>
-                  <tr>
-                    <td><?= $food["id"] ?></td>
-                    <td><img src="assets/images/<?= $food["image"] ?>" alt="IMG" width="50px"></td>
-                    <td><?= $food["name"] ?></td>
-                    <td><?= number_format($food["price"]) ?> VNĐ</td>
-                    <td><?= getNameDanhMuc($food["id_dm"])["name"] ?></td>
-                    <td class="box-active">
-                      <a href="?act=deleteFood&id=<?= $food["id"] ?>"
-                        onclick=" return confirm('Bạn có chắc muốn xóa không')"><button type="button"
-                          class="btn rounded btn-primary">
-                          <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor"
-                            class="bi bi-trash" viewBox="0 0 16 16">
-                            <path
-                              d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6Z">
-                            </path>
-                            <path
-                              d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1ZM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118ZM2.5 3h11V2h-11v1Z">
+                    <tr>
+                      <td><?= $food["id"] ?></td>
+                      <td><img src="assets/images/<?= $food["image"] ?>" alt="IMG" width="50px"></td>
+                      <td><?= $food["name"] ?></td>
+                      <td><?= number_format($food["price"]) ?> VNĐ</td>
+                      <td><?= getNameDanhMuc($food["id_dm"])["name"] ?></td>
+                      <td class="box-active">
+                        <a href="?act=deleteFood&id=<?= $food["id"] ?>" onclick=" return confirm('Bạn có chắc muốn xóa không')"><button type="button" class="btn rounded btn-primary">
+                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash" viewBox="0 0 16 16">
+                              <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6Z">
+                              </path>
+                              <path d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1ZM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118ZM2.5 3h11V2h-11v1Z">
+                              </path>
+                            </svg>
+                          </button>
+                        </a>
+                        <a href="?act=editFood&id=<?= $food["id"] ?>" class="btn btn-success rounded">
+                          <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-pen" viewBox="0 0 16 16">
+                            <path d="m13.498.795.149-.149a1.207 1.207 0 1 1 1.707 1.708l-.149.148a1.5 1.5 0 0 1-.059 2.059L4.854 14.854a.5.5 0 0 1-.233.131l-4 1a.5.5 0 0 1-.606-.606l1-4a.5.5 0 0 1 .131-.232l9.642-9.642a.5.5 0 0 0-.642.056L6.854 4.854a.5.5 0 1 1-.708-.708L9.44.854A1.5 1.5 0 0 1 11.5.796a1.5 1.5 0 0 1 1.998-.001zm-.644.766a.5.5 0 0 0-.707 0L1.95 11.756l-.764 3.057 3.057-.764L14.44 3.854a.5.5 0 0 0 0-.708l-1.585-1.585z">
                             </path>
                           </svg>
-                        </button>
-                      </a>
-                      <a href="?act=editFood&id=<?= $food["id"] ?>" class="btn btn-success rounded">
-                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor"
-                          class="bi bi-pen" viewBox="0 0 16 16">
-                          <path
-                            d="m13.498.795.149-.149a1.207 1.207 0 1 1 1.707 1.708l-.149.148a1.5 1.5 0 0 1-.059 2.059L4.854 14.854a.5.5 0 0 1-.233.131l-4 1a.5.5 0 0 1-.606-.606l1-4a.5.5 0 0 1 .131-.232l9.642-9.642a.5.5 0 0 0-.642.056L6.854 4.854a.5.5 0 1 1-.708-.708L9.44.854A1.5 1.5 0 0 1 11.5.796a1.5 1.5 0 0 1 1.998-.001zm-.644.766a.5.5 0 0 0-.707 0L1.95 11.756l-.764 3.057 3.057-.764L14.44 3.854a.5.5 0 0 0 0-.708l-1.585-1.585z">
-                          </path>
-                        </svg>
-                      </a>
-                    </td>
-                  </tr>
+                        </a>
+                      </td>
+                    </tr>
                   <?php endforeach; ?>
                 </tbody>
               </table>
